Modify warden to not store in session
Previously, GDTA was storing the authenticated user in the session. This
is likely not what was desired as evidenced by issue #7.

Therefore, this commit changes warden to only sign in a user for that
request. Since the user is not stored in session, the user has to be
pulled from warden (rather than the warden through the session).